requirements.txt
azure-functions
azure-cognitiveservices-speech
azure-cognitiveservices-language-luis
requests